修改&&新增 1.改__int128_t为long long 2.引入了lua虚拟机作为插件的执行环境

+++ editor/editorbase.cpp
@@ -1,13 +1,13 @@
 #include "editorbase.h"
 
-EditorBase::EditorBase(QString title,__int128_t id,
+EditorBase::EditorBase(QString title,long long id,
                        ProgramLanguageSupport language){
     this->title = &title;
     this->id = &id;
     this->language = &language;
 }
 
-EditorBase::EditorBase(QString title,__int128_t id){
+EditorBase::EditorBase(QString title,long long id){
     this->title = &title;
     this->id = &id;
 }
