+++ editor/editormanager.cpp
@@ -5,12 +5,12 @@ EditorManager::EditorManager(){
     editorVector.push_back(new EditorObject());
 }
 
-EditorObject* EditorManager::getEditor(__int128_t index){
+EditorObject* EditorManager::getEditor(long long index){
     return editorVector[index];
 }
 
 EditorManager::~EditorManager(){
-    for (__int128_t i = 0; i < editorVector.size(); i++)
+    for (long long i = 0; i < editorVector.size(); i++)
     {
         free(editorVector[i]);
     }
